// src/types/types.ts
export interface MyFormData {
  firstname: string;
  middlename: string;
  lastname: string;
  email: string;
  password: string;
  confirmPassword: string;
}

export interface admindata {
  _id: string;
  username: string;
  email: string;
  role: string;
  createdAt: string;
}

export interface tableData {
  _id: string;
  username: string;
  firstname: string;
  middlename: string;
  lastname: string;
  email: string;
  password: string;
  confirmPassword: string;
}